backend/requirements.txt
fastapi==0.111.0
slowapi==0.1.9
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
httpx==0.27.0
websockets==12.0
python-multipart==0.0.9
python-dotenv==1.0.1